Moved QTree data structure to utils.
